Still want some work, but much closer to what I want in the ned
