+++ Assets/Scripts/Controllers/Sprites/InventorySpriteController.cs
@@ -31,6 +31,8 @@ public sealed class InventorySpriteController : BaseSpriteController<Inventory>
         }
     }
 
+
+
     public static SpriteRenderer SetSprite(GameObject inventoryGO, Inventory inventory, string sortingLayerName = "Inventory")
     {
         SpriteRenderer sr = inventoryGO.GetComponent<SpriteRenderer>();
