+++ Assets/Scripts/UI/InGameUI/MenuRight/SelectionInfoTextField.cs
@@ -10,6 +10,7 @@ using System;
 using System.Linq;
 using ProjectPorcupine.Entities;
 using ProjectPorcupine.Localization;
+using ProjectPorcupine.Mouse;
 using UnityEngine;
 using UnityEngine.UI;
