+++ Assets/Scripts/UI/Mouse.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 375303736567456499ab8422848845eb
+folderAsset: yes
+timeCreated: 1502694562
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
