+++ Assets/Scripts/UI/Mouse/IMouseHandler.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 4072c2627e44a9241b112d69328f2748
+timeCreated: 1502694676
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
