+++ Assets/Scripts/UI/Mouse/MouseOver.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 00b4ef092c1372d47877e9ddfe38dfa6
+folderAsset: yes
+timeCreated: 1502694600
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
