+++ Assets/Scripts/UI/Mouse/MouseOver/MouseOverFurnitureTypeText.cs
